refactor(GameLoader): GameLoader 디렉터리 이동
